move Input component to global components and make paths relative
