+++ src/app/[locale]/_components/DashboardInput.tsx
@@ -2,8 +2,8 @@
 import { Typography } from '@mui/material';
 import { messages } from 'lib/i18n/localization';
 import { FormattedMessage } from 'react-intl';
-import { ManualAasInput } from 'app/[locale]/viewer/_components/ManualAasInput';
-import { QrScanner } from 'app/[locale]/_components/QrScanner';
+import { ManualAasInput } from '../_components/ManualAasInput';
+import { QrScanner } from '../_components/QrScanner';
 import { handleSearchForAas } from 'lib/searchUtilActions/search';
 import { useRouter } from 'next/navigation';
 import { useAasState, useRegistryAasState } from 'components/contexts/CurrentAasContext';
