+++ src/app/[locale]/_components/ManualAasInput.tsx
